解决部分 -Wall -Wextra 命令下的编译器警告

+++ lib/suplib.h
@@ -14,10 +14,10 @@
 		char* pMemC = pMem;
 		char* pSubC = pSub;
 		size_t iRange = iMemSize - iSubSize;
-		for ( int i = 0; i <= iRange; i++ ) {
+		for ( size_t i = 0; i <= iRange; i++ ) {
 			char* pPos = &pMemC[i];
 			int bOK = TRUE;
-			for ( int j = 0; j < iSubSize; j++ ) {
+			for ( size_t j = 0; j < iSubSize; j++ ) {
 				if ( pPos[j] != pSubC[j] ) {
 					bOK = FALSE;
 					break;
